Split the review GUI Streamlit panels into editor, generation, and visualization modules.
