+++ isaaclab_arena_examples/agentic_environment_generation/review_gui/visualization_panel.py
@@ -1,0 +1,25 @@
+# Copyright (c) 2026, The Isaac Lab Arena Project Developers (https://github.com/isaac-sim/IsaacLab-Arena/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+import streamlit as st
+
+_IFRAME_HEIGHT_PX = 1100
+
+
+def render_visualization_panel() -> None:
+    """Embed the rendered dashboard HTML in the right-hand column."""
+    st.subheader("Visualization")
+    if not st.session_state.get("last_rendered_text", "").strip():
+        st.caption("Generate or enter valid YAML to see the visualization.")
+        return
+
+    st.caption("Updates automatically when the YAML is valid.")
+    st.components.v1.html(
+        st.session_state["rendered_html"],
+        height=_IFRAME_HEIGHT_PX,
+        scrolling=True,
+    )
